requirements.txt
GitPython==2.1.8
requests==2.18.4
